Added the customers endpoint implementation

+++ src/Plugins/Nop.Plugin.Api/Services/ICustomerApiService.cs
@@ -7,9 +7,9 @@ namespace Nop.Plugin.Api.Services
     public interface ICustomerApiService
     {
         IList<CustomerDto> GetCustomersDtos(string createdAtMin = "", string createdAtMax = "",
-          byte limit = Configurations.DefaultLimit, int page = 1, int sinceId = 0);
+          int limit = Configurations.DefaultLimit, int page = 1, int sinceId = 0);
 
         int GetCustomersCount();
-        IList<CustomerDto> Search(Dictionary<string, string> query = null, string order = "desc", int page = 1, byte limit = Configurations.DefaultLimit);
+        IList<CustomerDto> Search(Dictionary<string, string> query = null, string order = "desc", int page = 1, int limit = Configurations.DefaultLimit);
     }
 }
